+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## v1.0.2
 
-- Fix bug that would occur if Journal entries weren't in a folder of the same name as the adventure.
+- Fixed bug that would occur if Journal entries weren't in a folder of the same name as the adventure.
 - Delay calling `scenePackerReady` hook until after the canvas is ready.
+- Added `await window['scene-packer'].relinkJournalEntries('module-name');` static method
+    - Automatically goes through the Journals in the module's compendium packs and updates the references from the World version to the Compendium versions.
 
 ## v1.0.1